refactor: pascalcase components and reorganized folders
