--- src/app/toaster.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { Toaster as SonnerToaster } from "sonner";
-
-export function Toaster() {
-  return <SonnerToaster theme="system" />;
-}
